Switch to using Vite instead of esbuild directly.
