+++ fm_eventmanager/settings.py.devel
@@ -91,6 +91,7 @@ INSTALLED_APPS = [
     'allauth.account',
     'allauth.mfa',
     'registration',
+    'django_vite',
 ]
 
 MIDDLEWARE = [
@@ -263,5 +264,7 @@ ENVIRONMENT_NAME = "Production Server"
 ENVIRONMENT_COLOR = "#FF0000"
 ENVIRONMENT_TEXT_COLOR = "#00FF00"
 ENVIRONMENT_FLOAT = True
+
+DJANGO_VITE_STATIC_URL_PREFIX = "bundler"
 
 VALIDATE_TEMPLATES_IGNORE_APPS = ["allauth"]
