+++ fm_eventmanager/settings.py.docker
@@ -134,6 +134,7 @@ INSTALLED_APPS = [
     'nested_inline',
     'debug_toolbar',
     'django_prometheus',
+    'django_vite',
 ]
 
 MIDDLEWARE = [
@@ -421,5 +422,7 @@ ENVIRONMENT_FLOAT = os.getenv('ENVIRONMENT_FLOAT', False)
 
 PRINT_RENDERER = os.getenv('PRINT_RENDERER', 'wkhtmltopdf')
 GOTENBERG_HOST = os.getenv('GOTENBERG_HOST', None)
+
+DJANGO_VITE_STATIC_URL_PREFIX = "bundler"
 
 VALIDATE_TEMPLATES_IGNORE_APPS = ["allauth"]
