+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "django-nested-inline~=0.4.6",
     "django-prometheus~=2.4.1",
     "django-redis~=6.0.0",
+    "django-vite>=3.1.0",
     "django-widget-tweaks~=1.5.1",
     "freezegun>=1.4.0",
     "gotenberg-client~=0.13.1",
