+++ registration/frontend/src/admin/attendee-search/components/BadgeTableLoader.tsx
@@ -1,4 +1,4 @@
-import { Component, For } from "solid-js";
+import { type Component, For } from "solid-js";
 
 export const BadgeTableLoader: Component<{ count?: number }> = (props) => {
   return (
