+++ registration/frontend/vite.config.ts
@@ -1,0 +1,25 @@
+import { defineConfig } from "vite";
+import solid from "vite-plugin-solid";
+
+export default defineConfig({
+  base: "/static/bundler/",
+  plugins: [solid()],
+  build: {
+    manifest: "manifest.json",
+    emptyOutDir: true,
+    outDir: "../static/bundler/",
+    rollupOptions: {
+      input: {
+        admin: "src/entrypoints/admin.tsx",
+      },
+    },
+  },
+  css: {
+    preprocessorOptions: {
+      scss: {
+        quietDeps: true,
+        silenceDeprecations: ["if-function"],
+      },
+    },
+  },
+});
